Daño critico y sistema de habilidades añadidos
Se ha añadido un sistema de probabilidad de críticos el cual multiplica el daño x2 en todas las armas según la probabilidad.
Además se ha empezado a crear el sistema de habilidades con el cual se puede aumentar las estadísticas del juego. De momento se han añadido: Aumentar vida máxima, aumentar la recuperación de vida automática, aumentar la probabilidad de robo de salud, aumentar el daño, aumentar el daño para armas melee, aumentar el daño para armas a distancia, disminuir el tiempo de recargas de armas, disminuir el tiempo entre generaciones de cajas, multiplicador de obtención de calaveras por muerte de enemigo

+++ Assets/Scripts/Enemigos/VidaEnemigo.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class VidaEnemigo : MonoBehaviour
 {
-    public float saludMaxima = 100f;
+    public int saludMaxima = 100;
 
-    private float saludActual;
+    private int saludActual;
     private DropCalaveras dropObjeto; // Referencia al script DropObjeto
     private BarraExperiencia barraExp;
 
@@ -16,7 +16,7 @@ public class VidaEnemigo : MonoBehaviour
 
     }
 
-    public void RecibirDaño(float cantidad)
+    public void RecibirDaño(int cantidad)
     {
         saludActual -= cantidad;
         if (saludActual <= 0f)
@@ -41,12 +41,12 @@ public class VidaEnemigo : MonoBehaviour
         Destroy(gameObject);
     }
 
-    public float ObtenerSalud()
+    public int ObtenerSalud()
     {
         return saludActual;
     }
 
-    public float ObtenerSaludMaxima()
+    public int ObtenerSaludMaxima()
     {
         return saludMaxima;
     }
